Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    double num1,num2,num3;
-    double average;
+    int num1,num2,num3;
+    float average;
     scanf("%d,%d,%d", &num1, &num2, &num3);
     average = (num1 + num2 + num3) / 3.0 ;
     printf("Average: %.2f\n",average);
